fix: handle multi-line descriptions correctly
In case where multi-line descriptions were defined using nix-heredoc,
the rendered markdown would contain newlines and indentations in the
beginning and end without trimming.

With this fix, the indentation is handled correctly while maintaining the
structure of the description for cases where the multi-line description
is supposed to be a paragraph.

Closes #3

Signed-off-by: Chinmay D. Pai <chinmaydpai@gmail.com>