fix: handle with expressions for parsing options
Currently if an option or a set of options were defined using `with lib;`
expression, the parser would skip over the set and not generate any output.

This patch adds support for parsing NODE_WITH kind, which should fix the issue
for now until we figure out a better solution.

Signed-off-by: Chinmay D. Pai <chinmaydpai@gmail.com>